Make the map the home page
