+++ app/transcribe/page.tsx
@@ -1,0 +1,9 @@
+import { MetadataForm } from "@/components/MetadataForm"
+
+export default function IndexPage() {
+  return (
+    <section className="container grid items-center gap-6 pb-8 pt-6 md:py-10">
+      <MetadataForm />
+    </section>
+  )
+}
